+++ CHANGELOG.md
@@ -12,3 +12,4 @@
 * ADD utilisateur - mot de passe perdu (+ recaptcha), validation par email
 * ADD utilisateur - interface de connexion / déconnexion
 * ADD utilisateur - mon profil - gestion de mes modules
+* ADD dev - script de mise à jour (semi) automatique: ./scripts/upgrade.sh